+++ frontend/package.json
@@ -14,6 +14,8 @@
     "jwt-decode": "^4.0.0",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.5",
+    "react-calendar": "^5.1.0",
+    "react-datepicker": "^7.5.0",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
     "react-modal": "^3.16.1",
